requirements.txt
# Core ML
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0          # YOLOv8
transformers>=4.30.0         # DETR via HuggingFace
datasets>=2.12.0

# Computer Vision
opencv-python>=4.7.0
Pillow>=9.5.0
albumentations>=1.3.0        # Data augmentation

# Data & Utils
numpy>=1.24.0
pandas>=2.0.0
pycocotools>=2.0.6           # COCO-style evaluation (mAP)
scipy>=1.10.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0

# Experiment Tracking (optional but recommended)
wandb>=0.15.0

# Notebook
jupyter>=1.0.0
ipywidgets>=8.0.0